Modyfikacja biblioteki C++, usunięcie zbednych plików

+++ Libs/Dll_C++/Dll_C++.cpp
@@ -5,11 +5,9 @@ extern "C" {
 	__declspec(dllexport)
 		void test(float matrix[], float scalar, int matrixLength)
 	{
-		for (size_t i = 0; i < matrixLength; i++)
+		for (int i = 0; i < matrixLength; i++)
 		{
 			matrix[i] *= scalar;
 		}
-
-		//Sleep(500);
 	}
 }
